:recycle: fix settings connect to emulator
