+++ lib/entrypoint.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:Pilll/analytics.dart';
 import 'package:Pilll/components/atoms/color.dart';
 import 'package:Pilll/service/push_notification.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_analytics/observer.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
@@ -17,11 +19,18 @@ Future<void> entrypoint() async {
   initializeDateFormatting('ja_JP');
   await Firebase.initializeApp();
   // MEMO: FirebaseCrashlytics#recordFlutterError called dumpErrorToConsole in function.
+  connectToEmulator();
   FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterError;
   requestNotificationPermissions();
   runZonedGuarded(() {
     runApp(ProviderScope(child: App()));
   }, (error, stack) => FirebaseCrashlytics.instance.recordError(error, stack));
+}
+
+void connectToEmulator() {
+  final domain = Platform.isAndroid ? '10.0.2.2' : 'localhost';
+  FirebaseFirestore.instance.settings = Settings(
+      persistenceEnabled: false, host: '$domain:8080', sslEnabled: false);
 }
 
 class App extends StatelessWidget {
