+++ lib/util/environment.dart
@@ -8,6 +8,7 @@ abstract class Environment {
   static bool get isProduction => flavor == Flavor.PRODUCTION;
   static bool get isDevelopment =>
       flavor == Flavor.DEVELOP || flavor == Flavor.LOCAL;
+  static bool get isLocal => flavor == Flavor.LOCAL;
   static bool isTest = false;
   static Flavor flavor;
 }
